// src/shared/components/containers/ProductConfiguratorAddToCart/types.ts
import { ChangeEvent } from 'react';
import { WithStyles } from '@material-ui/core';
import { styles } from './styles';
import {
    IProductPropFullData,
    TProductAvailability,
    TProductQuantity,
    TProductSKU,
    TProductType
} from 'src/shared/interfaces/product';
import { TCartId } from 'src/shared/interfaces/cart';
import { ICartCreatePayload } from 'src/shared/services/Common/Cart/types';

export interface ProductConfiguratorAddToCartProps extends WithStyles<typeof styles> {
    product: IProductPropFullData | null;
    productType: TProductType | null;
    sku: TProductSKU | null;
    isUserLoggedIn: boolean;
    addItemToCart: Function;
    addItemGuestCart: Function;
    createCartAndAddItem: Function;
    cartCreated: boolean;
    cartId: TCartId;
    payloadForCreateCart: ICartCreatePayload;
    anonymId: string;
}

export interface ProductConfiguratorAddToCartState {
    quantitySelected: TProductQuantity;
    quantity: TProductQuantity | null;
    isBuyBtnDisabled: boolean;
    isProcessCartLoading: boolean;
    availability: TProductAvailability | null;
    sku: TProductSKU | null;
}

export interface IProductQuantityParams {
    onChangeHandler: (event: ChangeEvent<HTMLSelectElement>) => void;
    inputValue: ProductConfiguratorAddToCartState['quantitySelected'];
    quantity: ProductConfiguratorAddToCartState['quantity'];
    controlsGroupClassName?: string;
}